Command Select Button Interface
Commands can be entered into the input box by clicking buttons displayed at the bottom of the right hand column.
Commands that take a parameter fetch lists of that parameter type and display as buttons above the command buttons and the command buttons are disabled.
Help command button replaced by cancel, which clears the command input box and hides the parameter selection.