Input History
A configurable number of input entries are stored in a buffer.
ui_up and ui_down events for the input LineEdit node are overridden and search though the history.
Any text already entered left of the caret is used as a search term.
Text recalled from history after the caret is selected.
History entries that are identical to the currently recalled text are skipped.
Entering text reset the history position to most recent.